+++ app/routes/schedule.py
@@ -4,6 +4,11 @@ from datetime import datetime
 
 @app.route("/hackathon/<hid>/schedule")
 def schedule(hid):
+    user = session['user']
+    teamIds = db.get_document(hid, "metadata", "data")['teamIds']
+    if not hid.startswith(user) and user not in teamIds:
+        return abort(403)
+
     hacks = db.get(hid)
     meta = db.get_document(hacks['$id'], "metadata", "data")
 
@@ -32,17 +37,32 @@ def schedule(hid):
 
 @app.post("/hackathon/<hackathon_id>/schedule/delete/<schedule_id>")
 def delete_schedule_item(hackathon_id, schedule_id):
+    user = session['user']
+    teamIds = db.get_document(hackathon_id, "metadata", "data")['teamIds']
+    if not hackathon_id.startswith(user) and user not in teamIds:
+        return abort(403)
+
     db.delete_document(hackathon_id, "schedule", schedule_id)
     return redirect(f"/hackathon/{hackathon_id}/schedule")
 
 @app.post("/hackathon/<hackathon_id>/schedule/edit/<schedule_id>")
 def edit_schedule_item(hackathon_id, schedule_id):
+    user = session['user']
+    teamIds = db.get_document(hackathon_id, "metadata", "data")['teamIds']
+    if not hackathon_id.startswith(user) and user not in teamIds:
+        return abort(403)
+
     data = {k: v for k, v in request.form.items()}
     db.update_document(hackathon_id, "schedule", schedule_id, data)
     return redirect(f"/hackathon/{hackathon_id}/schedule")
 
 @app.post("/hackathon/<hackathon_id>/schedule/add")
 def add_schedule_item(hackathon_id):
+    user = session['user']
+    teamIds = db.get_document(hackathon_id, "metadata", "data")['teamIds']
+    if not hackathon_id.startswith(user) and user not in teamIds:
+        return abort(403)
+        
     data = {k: v for k, v in request.form.items()}
     db.create_document(hackathon_id, "schedule", "unique()", data)
     return redirect(f"/hackathon/{hackathon_id}/schedule")
